requirements.txt
# requirements.txt
xgboost==1.5.0
scikit-learn==0.24.2
numpy==1.19.5
cloudml-hypertune==0.1.0
google-cloud-aiplatform==1.7.0
google-cloud-storage==1.44.0
pandas